Update test1.Rmd
Questions:
There’s a dataset in the folder called, “Bee colony Census Data by County”  and we wanted help in merging the data sets. 
Also in the “Bee colony Census Data by County” data set, there’s a column named “value” and we’ve looked in the glossary but we couldn’t find any info on it. We were wondering if you knew what we could use the “value” column for or if it’s even worth using anything in the ”Bee colony Census Data by County” data set.